Refactor: migrate shared code to Promise.Common project
Replaced Promise.Lib with Promise.Common for all shared interfaces, models, and utilities. Updated namespaces, project references, global usings, and workflow paths accordingly. Moved NLog config and related files. Removed Promise.Lib from the solution. Updated all usages and solution files to reference Promise.Common, improving code organization and maintainability.
