+++ Promise.Common/GlobalSuppressions.cs
@@ -1,0 +1,1 @@
+using System.Diagnostics.CodeAnalysis;
