+++ Promise.Common/GlobalUsings.cs
@@ -1,0 +1,1 @@
+// Add global usings here that are widely used across Promise.Common
