+++ Promise.Common/Interfaces/IMyEnvironment.cs
@@ -1,0 +1,11 @@
+﻿namespace Promise.Common;
+
+public interface IMyEnvironment
+{
+    bool IsDevelopment();
+    bool IsProduction();
+    string GetEnvironment();
+    bool IsNative();
+    bool IsWeb();
+    bool IsCutOff();
+}
