+++ Promise.Common/Interfaces/ISettings.cs
@@ -1,0 +1,8 @@
+﻿namespace Promise.Common;
+
+public interface ISettings
+{
+#pragma warning disable CA1056
+    string ApiUrl { get; }
+#pragma warning restore CA1056
+}
