+++ Promise.Common/Models/ApiResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace Promise.Common.Models;
+
+public class ApiResponse
+{
+    public bool Success { get; set; }
+    public string? Error { get; set; }
+
+}
