+++ Promise.Common/Models/ApiResponseUser.cs
@@ -1,0 +1,7 @@
+﻿namespace Promise.Common.Models;
+
+public class ApiResponseUser : ApiResponse
+{
+    public long Id { get; set; }
+    public string? Login { get; set; }
+}
