+++ Promise.Common/Models/ApiResponseUserInfo.cs
@@ -1,0 +1,7 @@
+﻿namespace Promise.Common.Models;
+
+public class ApiResponseUserInfo : ApiResponseUser
+{
+    public long Balance { get; set; }
+    public long PromiseLimit { get; set; }
+}
