+++ Promise.Common/NLog.config
