+++ Promise.Comp/Components/Pages/CurrencySettings.razor
@@ -2,6 +2,7 @@
 @using System.Net.Http.Headers
 @using System.Text
 @using System.Text.Json
+@using Promise.Common.Models
 @inject AppState appState
 @inject NavigationManager navigationManager
 @inject ISettings settings
