+++ Promise.Comp/Components/Pages/Subscriptions.razor
@@ -3,7 +3,7 @@
 @using System.Text
 @using System.Net.Http.Headers
 @using Promise.Comp.Components.Shared
-@using Promise.Lib.Models
+@using Promise.Common.Models
 @inject AppState appState
 @inject NavigationManager navigationManager
 @inject ISettings settings
