+++ Promise.Comp/Components/Pages/TransactionsHistory.razor
@@ -2,6 +2,7 @@
 @using System.Text.Json
 @using System.Text
 @using System.Net.Http.Headers
+@using Promise.Common.Models
 @using Promise.Comp.Components.Shared
 @inject AppState appState
 @inject NavigationManager navigationManager
