+++ Promise.Comp/GlobalUsings.cs
@@ -1,1 +1,2 @@
-global using Promise.Lib;
+global using Promise.Common;
+global using Promise.Common.Models;
