--- Promise.Lib/GlobalUsings.cs
@@ -1,1 +1,0 @@
-// Add global usings here that are widely used across Promise.Lib
