--- Promise.Lib/Interfaces/IMyEnvironment.cs
@@ -1,11 +1,0 @@
-﻿namespace Promise.Lib;
-
-public interface IMyEnvironment
-{
-    bool IsDevelopment();
-    bool IsProduction();
-    string GetEnvironment();
-    bool IsNative();
-    bool IsWeb();
-    bool IsCutOff();
-}
