--- Promise.Lib/Interfaces/ISettings.cs
@@ -1,8 +1,0 @@
-﻿namespace Promise.Lib;
-
-public interface ISettings
-{
-#pragma warning disable CA1056
-    string ApiUrl { get; }
-#pragma warning restore CA1056
-}
