--- Promise.Lib/Models/ApiResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace Promise.Lib.Models;
-
-public class ApiResponse
-{
-    public bool Success { get; set; }
-    public string? Error { get; set; }
-
-}
