--- Promise.Lib/Models/ApiResponseUser.cs
@@ -1,7 +1,0 @@
-﻿namespace Promise.Lib.Models;
-
-public class ApiResponseUser : ApiResponse
-{
-    public long Id { get; set; }
-    public string? Login { get; set; }
-}
