--- Promise.Lib/Models/ApiResponseUserInfo.cs
@@ -1,7 +1,0 @@
-﻿namespace Promise.Lib.Models;
-
-public class ApiResponseUserInfo : ApiResponseUser
-{
-    public long Balance { get; set; }
-    public long PromiseLimit { get; set; }
-}
