--- Promise.Lib/Promise.Lib.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" />
-    <PackageReference Include="NLog" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="NLog.config">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>  
-  
-</Project>
-
