+++ Promise.Native/GlobalUsings.cs
@@ -1,2 +1,2 @@
-global using Promise.Lib;
+global using Promise.Common;
 global using Promise.Comp;
