--- Promise.sln.old
@@ -1,59 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 18
-VisualStudioVersion = 18.1.11312.151
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Promise.Native", "Promise.Native\Promise.Native.csproj", "{32A44641-EE06-445C-B514-32452300DA0A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Promise.Web", "Promise.Web\Promise.Web.csproj", "{985499BD-209B-46F8-9D97-39DD5C58A700}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Promise.Comp", "Promise.Comp\Promise.Comp.csproj", "{83A37629-941F-44BB-8D81-F51B4D785002}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Promise.Lib", "Promise.Lib\Promise.Lib.csproj", "{A2DD7608-793E-46E8-8D0A-5DE5F1E0E5DA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Promise.Api", "Promise.Api\Promise.Api.csproj", "{186E2A46-D4E6-4053-87FE-78E2948E88D4}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Deployment", "Deployment", "{12718B9F-EFCB-4ACA-BAB4-7B9B885B14C1}"
-	ProjectSection(SolutionItems) = preProject
-		info.md = info.md
-		.github\workflows\ios_deploy_promisenative.yml = .github\workflows\ios_deploy_promisenative.yml
-		.github\workflows\main_promiseapi.yml = .github\workflows\main_promiseapi.yml
-		.github\workflows\main_promisesite.yml = .github\workflows\main_promisesite.yml
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{32A44641-EE06-445C-B514-32452300DA0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{32A44641-EE06-445C-B514-32452300DA0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{32A44641-EE06-445C-B514-32452300DA0A}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{32A44641-EE06-445C-B514-32452300DA0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{32A44641-EE06-445C-B514-32452300DA0A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{32A44641-EE06-445C-B514-32452300DA0A}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{985499BD-209B-46F8-9D97-39DD5C58A700}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{985499BD-209B-46F8-9D97-39DD5C58A700}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{985499BD-209B-46F8-9D97-39DD5C58A700}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{985499BD-209B-46F8-9D97-39DD5C58A700}.Release|Any CPU.Build.0 = Release|Any CPU
-		{83A37629-941F-44BB-8D81-F51B4D785002}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{83A37629-941F-44BB-8D81-F51B4D785002}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{83A37629-941F-44BB-8D81-F51B4D785002}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{83A37629-941F-44BB-8D81-F51B4D785002}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A2DD7608-793E-46E8-8D0A-5DE5F1E0E5DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A2DD7608-793E-46E8-8D0A-5DE5F1E0E5DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A2DD7608-793E-46E8-8D0A-5DE5F1E0E5DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A2DD7608-793E-46E8-8D0A-5DE5F1E0E5DA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{186E2A46-D4E6-4053-87FE-78E2948E88D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{186E2A46-D4E6-4053-87FE-78E2948E88D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{186E2A46-D4E6-4053-87FE-78E2948E88D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{186E2A46-D4E6-4053-87FE-78E2948E88D4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CA961516-BF85-4334-9F75-ABF7385B0F8C}
-	EndGlobalSection
-EndGlobal
